2nd level and extra stuff
- I added a bool to the health system for enemies to respawn
- I created a prefab for level transitions
- Some small changes for how the power ups cooldowns work
- Lowerd the dificulty of the first level (what the teacher couldn't pass)
- Created a bool for GameOver, to let the player auto respawn

+++ Assets/Scripts/PowerUpCollectable.cs
@@ -4,7 +4,7 @@ public class PowerUpCollectable : MonoBehaviour
 {
     [SerializeField] private PlayerController.PowerUp powerUpType = PlayerController.PowerUp.Movement;
     [SerializeField] private float multilpierAmount = 0;
-    [SerializeField] private float cooldown = 10;
+    [SerializeField] private int cooldown = 10;
 
     void OnTriggerEnter(Collider other)
     {
